% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{applyCPO}
\alias{applyCPO}
\title{CPO Apply}
\usage{
applyCPO(cpo, task)
}
\arguments{
\item{cpo}{[\code{\link{CPO}}]\cr
The CPO representing the operation to perform.}

\item{task}{[\code{\link{Task}}]\cr
The task to operate on.}
}
\description{
The given transformation will be applied to the data in the given \code{link{Task}}.

See the preferred \code{\link{\%>>\%}} for more info.
}
\seealso{
Other CPO: \code{\link{CPOImputer}}, \code{\link{\%>>\%}},
  \code{\link{attachCPO}}, \code{\link{cpoApplyFun}},
  \code{\link{cpoAsNumeric}}, \code{\link{cpoCase}},
  \code{\link{cpoCollapseFact}},
  \code{\link{cpoDropConstants}},
  \code{\link{cpoDummyEncode}},
  \code{\link{cpoFilterFeatures}},
  \code{\link{cpoFixFactors}},
  \code{\link{cpoImpactEncodeClassif}},
  \code{\link{cpoImpactEncodeRegr}},
  \code{\link{cpoImpute}},
  \code{\link{cpoMissingIndicators}},
  \code{\link{cpoModelMatrix}}, \code{\link{cpoMultiplex}},
  \code{\link{cpoPca}}, \code{\link{cpoProbEncode}},
  \code{\link{cpoQuantileBinNumerics}},
  \code{\link{cpoScaleMaxAbs}},
  \code{\link{cpoScaleRange}}, \code{\link{cpoScale}},
  \code{\link{cpoSelect}}, \code{\link{cpoSpatialSign}},
  \code{\link{cpoWrap}}, \code{\link{getCPOAffect}},
  \code{\link{getCPOId}}, \code{\link{getCPOName}},
  \code{\link{getRetrafoState}},
  \code{\link{makeCPOExtended}}, \code{\link{makeCPO}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{pipeCPO}}, \code{\link{setCPOId}}
}
