% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_cbind.R
\name{cpoCbind}
\alias{cpoCbind}
\title{\dQuote{cbind} the Result of multiple CPOs}
\usage{
cpoCbind(..., .cpos = list())
}
\arguments{
\item{...}{[\code{CPO}]\cr
The CPOs to cbind. Named arguments will result in the respective
columns being prefixed with the name. This is highly recommended if there
is any chance of name collision otherwise. it is possible to use the same
name multiple times.}

\item{.cpos}{[\code{list} of \code{CPO}]\cr
Alternatively, give the CPOs to cbind as a list. Default is \code{list()}.}

\item{id}{[\code{character(1)}]\cr
id to use as prefix for the CPO's hyperparameters. this
must be used to avoid name clashes when composing two
CPOs of the same type, or with learners or other CPOS
with hyperparameters with clashing names.}

\item{export}{[\code{character}]\cr
Either a character vector indicating the parameters to
export as hyperparameters, or one of the special values
\dQuote{export.all} (export all parameters),
\dQuote{export.default} (export all parameters that are exported by default),
\dQuote{export.set} (export all parameters that were set during construction),
\dQuote{export.default.set} (export the intersection of the \dQuote{default} and \dQuote{set} parameters),
\dQuote{export.unset} (export all parameters that were \emph{not} set during construction) or
\dQuote{export.default.unset} (export the intersection of the \dQuote{default} and \dQuote{unset} parameters).
Default is \dQuote{export.default}.}

\item{affect.type}{[\code{character} | \code{NULL}]\cr
Type of columns to affect. A subset of \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}, \dQuote{other}, or \code{NULL}
to not match by column type. Default is \code{NULL}.}

\item{affect.index}{[\code{numeric}]\cr
Indices of feature columns to affect. The order of indices given is respected. Target column indices are not counted
(since target columns are always included). Default is \code{integer(0)}.}

\item{affect.names}{[\code{character}]\cr
Feature names of feature columns to affect. The order of names given is respected. Default is \code{character(0)}.}

\item{affect.pattern}{[\code{character(1)} | \code{NULL}]\cr
\code{\link[base]{grep}} pattern to match feature names by. Default is \code{NULL} (no pattern matching)}

\item{affect.invert}{[\code{logical(1)}]\cr
Whether to affect all features \emph{not} matched by other \code{affect.*} parameters.}

\item{affect.pattern.ignore.case}{[\code{logical(1)}]\cr
Ignore case when matching features with \code{affect.pattern}; see \code{\link[base]{grep}}. Default is \code{FALSE}.}

\item{affect.pattern.perl}{[\code{logical(1)}]\cr
Use Perl-style regular expressions for \code{affect.pattern}; see \code{\link[base]{grep}}. Default is \code{FALSE}.}

\item{affect.pattern.fixed}{[\code{logical(1)}]\cr
Use fixed matching instead of regular expressions for \code{affect.pattern}; see \code{\link[base]{grep}}. Default is \code{FALSE}.}
}
\value{
[\code{\link{CPO}}].
}
\description{
This is a \code{\link{CPOConstructor}} to be used to create a
\code{\link{CPO}}. It is called like any R function and returns
the created \code{\link{CPO}}.

Build a CPO that represents the operations of its input parameters,
performed in parallel and put together column wise.

For example, to construct a \code{\link{Task}} that contains the original
data, as well as the data after scaling, one could do

\code{task \%>>\% cpoCbind(NULLCPO, cpoScale())}

The result of \code{cpoCbind} is itself a CPO which exports its constituents'
hyperparameters. CPOs with the same type / ID get combined automatically.
To get networks, e.g. of the form
\preformatted{
      ,-C--E-.
     /    /   \
A---B ---D-----F---G
}
one coul use the code
\preformatted{
initcpo = A \%>>\% B
route1 = initcpo \%>>\% D
route2 = cpoCbind(route1, initcpo \%>>\% C) \%>>\% E
result = cpoCbind(route1, route2) \%>>\% F \%>>\% G
}

cpoCbind finds common paths among its arguments and combines them into one operation.
This saves computation and makes it possible for one exported hyperparameter to
influence multiple of \code{cpoCbind}'s inputs. However, if you want to use the same
operation with different parameters on different parts of \code{cpoCbind} input,
you must give these operations different IDs. If CPOs that could represent an identical CPO,
with the same IDs (or both with IDs absent) but different parameter settings or different
parameter exportations occur, an error will be thrown.
}
\section{General CPO info}{

This function creates a CPO object, which can be applied to
\code{\link{Task}}s, \code{data.frame}s, \code{link{Learner}}s
and other CPO objects using the \code{\link{\%>>\%}} operator.

The parameters of this object can be changed after creation
using the function \code{\link{setHyperPars}}. The other
hyper-parameter manipulating functins, \code{\link{getHyperPars}}
and \code{\link{getParamSet}} similarly work as one expects.

If the \dQuote{id} parameter is given, the hyperparameters
will have this id as aprefix; this will, however, not change
the parameters of the creator function.
}

\section{Calling a \code{\link{CPOConstructor}}}{

CPO constructor functions are called with optional values of parameters, and additional \dQuote{special} optional values.
The special optional values are the \code{id} parameter, and the \code{affect.*} parameters. The \code{affect.*} parameters
enable the user to control which subset of a given dataset is affected. If no \code{affect.*} parameters are given, all
data features are affected by default.
}

\seealso{
Other CPOs: \code{\link{cpoApplyFun}},
  \code{\link{cpoAsNumeric}}, \code{\link{cpoCase}},
  \code{\link{cpoCollapseFact}},
  \code{\link{cpoDropConstants}},
  \code{\link{cpoDummyEncode}},
  \code{\link{cpoFilterAnova}},
  \code{\link{cpoFilterCarscore}},
  \code{\link{cpoFilterChiSquared}},
  \code{\link{cpoFilterFeatures}},
  \code{\link{cpoFilterGainRatio}},
  \code{\link{cpoFilterInformationGain}},
  \code{\link{cpoFilterKruskal}},
  \code{\link{cpoFilterLinearCorrelation}},
  \code{\link{cpoFilterMrmr}}, \code{\link{cpoFilterOneR}},
  \code{\link{cpoFilterPermutationImportance}},
  \code{\link{cpoFilterRankCorrelation}},
  \code{\link{cpoFilterRelief}},
  \code{\link{cpoFilterRfCImportance}},
  \code{\link{cpoFilterRfImportance}},
  \code{\link{cpoFilterRfSRCImportance}},
  \code{\link{cpoFilterRfSRCMinDepth}},
  \code{\link{cpoFilterSymmetricalUncertainty}},
  \code{\link{cpoFilterUnivariate}},
  \code{\link{cpoFilterVariance}},
  \code{\link{cpoFixFactors}},
  \code{\link{cpoImpactEncodeClassif}},
  \code{\link{cpoImpactEncodeRegr}},
  \code{\link{cpoImpute}},
  \code{\link{cpoMissingIndicators}},
  \code{\link{cpoModelMatrix}}, \code{\link{cpoMultiplex}},
  \code{\link{cpoPca}}, \code{\link{cpoProbEncode}},
  \code{\link{cpoQuantileBinNumerics}},
  \code{\link{cpoScaleMaxAbs}},
  \code{\link{cpoScaleRange}}, \code{\link{cpoScale}},
  \code{\link{cpoSelect}}, \code{\link{cpoSpatialSign}},
  \code{\link{cpoWrap}}

Other special CPOs: \code{\link{cpoCase}},
  \code{\link{cpoMultiplex}}, \code{\link{cpoWrap}}
}
