% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_filterFeatures.R
\name{cpoFilterFeatures}
\alias{cpoFilterFeatures}
\title{Filter features by thresholding filter values.}
\usage{
cpoFilterFeatures(method = "randomForestSRC.rfsrc", fval = NULL,
  perc = NULL, abs = NULL, threshold = NULL, filter.args = list(),
  id = NULL, export = "export.default", affect.type = NULL,
  affect.index = integer(0), affect.names = character(0),
  affect.pattern = NULL, affect.invert = FALSE,
  affect.pattern.ignore.case = FALSE, affect.pattern.perl = FALSE,
  affect.pattern.fixed = FALSE)
}
\arguments{
\item{method}{[\code{character(1)}]\cr
See \code{\link{listFilterMethods}}.
Default is \dQuote{randomForestSRC.rfsrc}.}

\item{fval}{[\code{\link{FilterValues}}]\cr
Result of \code{\link{generateFilterValuesData}}.
If you pass this, the filter values in the object are used for feature filtering.
\code{method} and \code{...} are ignored then.
Default is \code{NULL} and not used.}

\item{perc}{[\code{numeric(1)}]\cr
If set, select \code{perc}*100 top scoring features.
Mutually exclusive with arguments \code{abs} and \code{threshold}.}

\item{abs}{[\code{numeric(1)}]\cr
If set, select \code{abs} top scoring features.
Mutually exclusive with arguments \code{perc} and \code{threshold}.}

\item{threshold}{[\code{numeric(1)}]\cr
If set, select features whose score exceeds \code{threshold}.
Mutually exclusive with arguments \code{perc} and \code{abs}.}

\item{filter.args}{[\code{list}]\cr
Passed down to selected filter method. Default is \code{list()}.}

\item{id}{[\code{character(1)}]\cr
id to use as prefix for the CPO's hyperparameters. this
must be used to avoid name clashes when composing two
CPOs of the same type, or with learners or other CPOS
with hyperparameters with clashing names.}

\item{export}{[\code{character}]\cr
Either a character vector indicating the parameters to
export as hyperparameters, or one of the special values
\dQuote{export.all} (export all parameters),
\dQuote{export.default} (export all parameters that are exported by default),
\dQuote{export.set} (export all parameters that were set during construction),
\dQuote{export.default.set} (export the intersection of the \dQuote{default} and \dQuote{set} parameters),
\dQuote{export.unset} (export all parameters that were \emph{not} set during construction) or
\dQuote{export.default.unset} (export the intersection of the \dQuote{default} and \dQuote{unset} parameters).
Default is \dQuote{export.default}.}

\item{affect.type}{[\code{character} | \code{NULL}]\cr
Type of columns to affect. A subset of \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}, \dQuote{other}, or \code{NULL}
to not match by column type. Default is \code{NULL}.}

\item{affect.index}{[\code{numeric}]\cr
Indices of feature columns to affect. The order of indices given is respected. Default is \code{integer(0)}.}

\item{affect.names}{[\code{character}]\cr
Feature names of feature columns to affect. The order of names given is respected. Default is \code{character(0)}.}

\item{affect.pattern}{[\code{character(1)} | \code{NULL}]\cr
\code{\link[base]{grep}} pattern to match feature names by. Default is \code{NULL} (no pattern matching)}

\item{affect.invert}{[\code{logical(1)}]\cr
Whether to affect all features \emph{not} matched by other \code{affect.*} parameters.}

\item{affect.pattern.ignore.case}{[\code{logical(1)}]\cr
Ignore case when matching features with \code{affect.pattern}; see \code{\link[base]{grep}}. Default is \code{FALSE}.}

\item{affect.pattern.perl}{[\code{logical(1)}]\cr
Use Perl-style regular expressions for \code{affect.pattern}; see \code{\link[base]{grep}}. Default is \code{FALSE}.}

\item{affect.pattern.fixed}{[\code{logical(1)}]\cr
Use fixed matching instead of regular expressions for \code{affect.pattern}; see \code{\link[base]{grep}}. Default is \code{FALSE}.}
}
\description{
First, calls \code{\link{generateFilterValuesData}}.
Features are then selected via \code{select} and \code{val}.

CPO constructor functions are called with optional values of parameters, and additional \dQuote{special} optional values.
  The special optional values are the \code{id} parameter, and the \code{affect.*} parameters. The \code{affect.*} parameters
  enable the user to control which subset of a given dataset is affected. If no \code{affect.*} parameters are given, all
  data features are affected by default.
}
\section{General CPO Info}{
NA
}

\seealso{
Other CPO: \code{\link{CPOImputer}}, \code{\link{\%>>\%}},
  \code{\link{applyCPO}}, \code{\link{attachCPO}},
  \code{\link{cpoApplyFun}}, \code{\link{cpoAsNumeric}},
  \code{\link{cpoCase}}, \code{\link{cpoCollapseFact}},
  \code{\link{cpoDropConstants}},
  \code{\link{cpoDummyEncode}},
  \code{\link{cpoFixFactors}},
  \code{\link{cpoImpactEncodeClassif}},
  \code{\link{cpoImpactEncodeRegr}},
  \code{\link{cpoImpute}},
  \code{\link{cpoMissingIndicators}},
  \code{\link{cpoModelMatrix}}, \code{\link{cpoMultiplex}},
  \code{\link{cpoPca}}, \code{\link{cpoProbEncode}},
  \code{\link{cpoQuantileBinNumerics}},
  \code{\link{cpoScaleMaxAbs}},
  \code{\link{cpoScaleRange}}, \code{\link{cpoScale}},
  \code{\link{cpoSelect}}, \code{\link{cpoSpatialSign}},
  \code{\link{cpoWrap}}, \code{\link{getCPOAffect}},
  \code{\link{getCPOId}}, \code{\link{getCPOName}},
  \code{\link{getRetrafoState}},
  \code{\link{makeCPOExtended}}, \code{\link{makeCPO}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{pipeCPO}}, \code{\link{setCPOId}}
}
