% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_regrResiduals.R
\name{cpoRegrResiduals}
\alias{cpoRegrResiduals}
\title{Train a Model on a Task and Return the Residual Task}
\usage{
cpoRegrResiduals(learner, predict.se = FALSE, crr.train.residuals = "plain",
  crr.resampling = cv5, id, export = "export.default", affect.type = NULL,
  affect.index = integer(0), affect.names = character(0),
  affect.pattern = NULL, affect.invert = FALSE,
  affect.pattern.ignore.case = FALSE, affect.pattern.perl = FALSE,
  affect.pattern.fixed = FALSE)
}
\arguments{
\item{learner}{[\code{character(1)} | \code{\link[mlr:makeLearner]{Learner}}]\cr
A regression \code{\link[mlr:makeLearner]{Learner}}, or a \code{character(1)} identifying a
\code{\link[mlr:makeLearner]{Learner}} to be constructed.}

\item{predict.se}{[\code{logical(1)}]\cr
Whether to fit the model with \dQuote{se} predict type. This enables the resulting
\code{\link{CPOInverter}} to be used for \code{property.type == "se"} inversion.
Default is \code{FALSE}.}

\item{crr.train.residuals}{[\code{character(1)}]\cr
What residuals to use for training (i.e. initial transformation). One of \dQuote{resample}, \dQuote{oob},
\dQuote{plain}. If \dQuote{resample} is given, the out-of-resampling-fold predictions are used when resampling
according to the \code{resampling} parameter. If \dQuote{oob} is used, the \code{\link[mlr:makeLearner]{Learner}} must
have the \dQuote{oobpreds} property; the out-of-bag predictions are then used. If \code{train.residuals} is
\dQuote{plain}, the simple regression residuals are used. \dQuote{plain} may offer slightly worse performance
than the alternatives, but few \code{mlr} \code{\link[mlr:makeLearner]{Learners}} support \dQuote{oobpreds}, and
\dQuote{resample} can come at a considerable run time penalty. Default is \dQuote{plain}.}

\item{crr.resampling}{[\code{\link[mlr:makeResampleDesc]{ResampleDesc}} | \code{\link[mlr:makeResampleInstance]{ResampleInstance}}]\cr
What resampling to use when \code{train.residuals} is \dQuote{resample}; otherwise has no effect.
The \code{$predict} slot of the resample description will be ignored and set to \code{test}.
If a data point is predicted by multiple resampling folds, the average residual is used. If a data
point is not predicted by any resampling fold, the \dQuote{plain} residual is used for that one.
Default is \code{cv5}.}

\item{id}{[\code{character(1)}]\cr
id to use as prefix for the CPO's hyperparameters. this
must be used to avoid name clashes when composing two
CPOs of the same type, or with learners or other CPOS
with hyperparameters with clashing names.}

\item{export}{[\code{character}]\cr
Either a character vector indicating the parameters to
export as hyperparameters, or one of the special values
\dQuote{export.all} (export all parameters),
\dQuote{export.default} (export all parameters that are exported by default),
\dQuote{export.set} (export all parameters that were set during construction),
\dQuote{export.default.set} (export the intersection of the \dQuote{default} and \dQuote{set} parameters),
\dQuote{export.unset} (export all parameters that were \emph{not} set during construction) or
\dQuote{export.default.unset} (export the intersection of the \dQuote{default} and \dQuote{unset} parameters).
Default is \dQuote{export.default}.}

\item{affect.type}{[\code{character} | \code{NULL}]\cr
Type of columns to affect. A subset of \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}, \dQuote{other}, or \code{NULL}
to not match by column type. Default is \code{NULL}.}

\item{affect.index}{[\code{numeric}]\cr
Indices of feature columns to affect. The order of indices given is respected. Target column indices are not counted
(since target columns are always included). Default is \code{integer(0)}.}

\item{affect.names}{[\code{character}]\cr
Feature names of feature columns to affect. The order of names given is respected. Default is \code{character(0)}.}

\item{affect.pattern}{[\code{character(1)} | \code{NULL}]\cr
\code{\link[base]{grep}} pattern to match feature names by. Default is \code{NULL} (no pattern matching)}

\item{affect.invert}{[\code{logical(1)}]\cr
Whether to affect all features \emph{not} matched by other \code{affect.*} parameters.}

\item{affect.pattern.ignore.case}{[\code{logical(1)}]\cr
Ignore case when matching features with \code{affect.pattern}; see \code{\link[base]{grep}}. Default is \code{FALSE}.}

\item{affect.pattern.perl}{[\code{logical(1)}]\cr
Use Perl-style regular expressions for \code{affect.pattern}; see \code{\link[base]{grep}}. Default is \code{FALSE}.}

\item{affect.pattern.fixed}{[\code{logical(1)}]\cr
Use fixed matching instead of regular expressions for \code{affect.pattern}; see \code{\link[base]{grep}}. Default is \code{FALSE}.}
}
\value{
[\code{\link{CPO}}].
}
\description{
This is a \code{\link{CPOConstructor}} to be used to create a
\code{\link{CPO}}. It is called like any R function and returns
the created \code{\link{CPO}}.

Given a regression learner, this \code{\link{CPO}} fits the learner to a
regression \code{\link[mlr]{Task}} and replaces the regression target with
the residuals--the differences of the target values and the model's predictions--of the model.

For inversion, the predictions of the model for the prediction data are
added to the predictions to be inverted.

If \code{predict.se} is \code{TRUE}, \code{property.type == "se"} inversion can also
be performed. In that case, the \code{se} of the incoming prediction and the \code{se}
of the internal model are assumed to be independently distributed, and the resulting
\code{se} is the pythagorean sum of the \code{se}s.
}
\section{CPOTrained State}{

The \code{CPORetrafo} state's \code{$control} slot is the \code{\link[mlr:makeWrappedModel]{WrappedModel}}
created when training the \code{learner} on the given data.

The \code{CPOInverter} state's \code{$control} slot is a \code{data.frame} of the \dQuote{response} and
(if \code{predict.se} is \code{TRUE}) \dQuote{se} columns of the prediction done by the model on the data.
}

\section{General CPO info}{

This function creates a CPO object, which can be applied to
\code{\link{Task}}s, \code{data.frame}s, \code{link{Learner}}s
and other CPO objects using the \code{\link{\%>>\%}} operator.

The parameters of this object can be changed after creation
using the function \code{\link{setHyperPars}}. The other
hyper-parameter manipulating functins, \code{\link{getHyperPars}}
and \code{\link{getParamSet}} similarly work as one expects.

If the \dQuote{id} parameter is given, the hyperparameters
will have this id as aprefix; this will, however, not change
the parameters of the creator function.
}

\section{Calling a \code{\link{CPOConstructor}}}{

CPO constructor functions are called with optional values of parameters, and additional \dQuote{special} optional values.
The special optional values are the \code{id} parameter, and the \code{affect.*} parameters. The \code{affect.*} parameters
enable the user to control which subset of a given dataset is affected. If no \code{affect.*} parameters are given, all
data features are affected by default.
}

\seealso{
Other CPOs: \code{\link{cpoApplyFunRegrTarget}},
  \code{\link{cpoApplyFun}}, \code{\link{cpoAsNumeric}},
  \code{\link{cpoCache}}, \code{\link{cpoCbind}},
  \code{\link{cpoCollapseFact}},
  \code{\link{cpoDropConstants}},
  \code{\link{cpoDummyEncode}},
  \code{\link{cpoFilterAnova}},
  \code{\link{cpoFilterCarscore}},
  \code{\link{cpoFilterChiSquared}},
  \code{\link{cpoFilterFeatures}},
  \code{\link{cpoFilterGainRatio}},
  \code{\link{cpoFilterInformationGain}},
  \code{\link{cpoFilterKruskal}},
  \code{\link{cpoFilterLinearCorrelation}},
  \code{\link{cpoFilterMrmr}}, \code{\link{cpoFilterOneR}},
  \code{\link{cpoFilterPermutationImportance}},
  \code{\link{cpoFilterRankCorrelation}},
  \code{\link{cpoFilterRelief}},
  \code{\link{cpoFilterRfCImportance}},
  \code{\link{cpoFilterRfImportance}},
  \code{\link{cpoFilterRfSRCImportance}},
  \code{\link{cpoFilterRfSRCMinDepth}},
  \code{\link{cpoFilterSymmetricalUncertainty}},
  \code{\link{cpoFilterUnivariate}},
  \code{\link{cpoFilterVariance}},
  \code{\link{cpoFixFactors}}, \code{\link{cpoIca}},
  \code{\link{cpoImpactEncodeClassif}},
  \code{\link{cpoImpactEncodeRegr}},
  \code{\link{cpoImputeConstant}},
  \code{\link{cpoImputeHist}},
  \code{\link{cpoImputeLearner}},
  \code{\link{cpoImputeMax}}, \code{\link{cpoImputeMean}},
  \code{\link{cpoImputeMedian}},
  \code{\link{cpoImputeMin}}, \code{\link{cpoImputeMode}},
  \code{\link{cpoImputeNormal}},
  \code{\link{cpoImputeUniform}}, \code{\link{cpoImpute}},
  \code{\link{cpoLogTrafoRegr}}, \code{\link{cpoMakeCols}},
  \code{\link{cpoMissingIndicators}},
  \code{\link{cpoModelMatrix}},
  \code{\link{cpoOversample}}, \code{\link{cpoPca}},
  \code{\link{cpoProbEncode}},
  \code{\link{cpoQuantileBinNumerics}},
  \code{\link{cpoResponseFromSE}}, \code{\link{cpoSample}},
  \code{\link{cpoScaleMaxAbs}},
  \code{\link{cpoScaleRange}}, \code{\link{cpoScale}},
  \code{\link{cpoSelect}}, \code{\link{cpoSmote}},
  \code{\link{cpoSpatialSign}},
  \code{\link{cpoTransformParams}}, \code{\link{cpoWrap}},
  \code{\link{makeCPOCase}}, \code{\link{makeCPOMultiplex}}
}
