% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getCPOAffect}
\alias{getCPOAffect}
\title{Get the Selection Arguments for affected CPOs}
\usage{
getCPOAffect(cpo, drop.defaults = TRUE)
}
\arguments{
\item{cpo}{[\code{CPO}]\cr
The CPO.}

\item{drop.defaults}{[\code{logical(1)}]\cr
Whether to only return the arguments that deviate from the default.
Default is \code{TRUE}.}
}
\description{
Get the \code{affected.*} arguments from when the CPO was constructed.
}
\seealso{
Other CPO: \code{\link{CPOImputer}}, \code{\link{CPO}},
  \code{\link{\%>>\%}}, \code{\link{applyCPO}},
  \code{\link{attachCPO}}, \code{\link{cpoApplyFun}},
  \code{\link{cpoAsNumeric}}, \code{\link{cpoCase}},
  \code{\link{cpoCollapseFact}},
  \code{\link{cpoDropConstants}},
  \code{\link{cpoDummyEncode}},
  \code{\link{cpoFilterFeatures}},
  \code{\link{cpoFixFactors}},
  \code{\link{cpoImpactEncodeClassif}},
  \code{\link{cpoImpactEncodeRegr}},
  \code{\link{cpoImpute}},
  \code{\link{cpoMissingIndicators}},
  \code{\link{cpoModelMatrix}}, \code{\link{cpoMultiplex}},
  \code{\link{cpoPca}}, \code{\link{cpoProbEncode}},
  \code{\link{cpoQuantileBinNumerics}},
  \code{\link{cpoScaleMaxAbs}},
  \code{\link{cpoScaleRange}}, \code{\link{cpoScale}},
  \code{\link{cpoSelect}}, \code{\link{cpoSpatialSign}},
  \code{\link{cpoWrap}}, \code{\link{getCPOId}},
  \code{\link{getCPOName}}, \code{\link{getRetrafoState}},
  \code{\link{inverter<-}}, \code{\link{inverter}},
  \code{\link{makeCPOExtended}}, \code{\link{makeCPO}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{pipeCPO}}, \code{\link{retrafo<-}},
  \code{\link{retrafo}}, \code{\link{setCPOId}}
}
