% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{getCPOId}
\alias{getCPOId}
\title{Get the ID of a CPO Object}
\usage{
getCPOId(cpo)
}
\arguments{
\item{cpo}{[\code{\link{CPO}}]\cr
The cpo.}
}
\value{
[\code{character(1)}] the CPO's id.
}
\description{
Gets the \emph{id} of a \code{\link{CPO}}. The id can be set
during construction by a \code{\link{CPOConstructor}}
using the \code{id} parameter, or with \code{\link{setCPOId}}.

The exported hyperparameters of a CPO all have the
id as prefix. This makes it possible to
compose CPOs that have clashing parameter names.
}
\seealso{
Other getters and setters: \code{\link{CPO}},
  \code{\link{getCPOAffect}}, \code{\link{getCPOClass}},
  \code{\link{getCPOConstructor}},
  \code{\link{getCPOName}},
  \code{\link{getCPOOperatingType}},
  \code{\link{getCPOPredictType}},
  \code{\link{getCPOProperties}},
  \code{\link{getCPOTrainedCPO}},
  \code{\link{getCPOTrainedCapability}},
  \code{\link{setCPOId}}

Other CPO ID related: \code{\link{setCPOId}}
}
\concept{CPO ID related}
\concept{getters and setters}
