% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.R
\name{getLearnerBare}
\alias{getLearnerBare}
\title{Get the Learner with the CPOs Removed}
\usage{
getLearnerBare(learner)
}
\arguments{
\item{learner}{[\code{\link[mlr:makeLearner]{Learner}}]\cr
The learner to strip.}
}
\value{
[\code{\link[mlr:makeLearner]{Learner}}]. The learner without attached CPOs.
}
\description{
Get the bare \code{\link[mlr:makeLearner]{Learner}} without the \code{\link{CPO}}s that were previously added.

It is still possible for the result to be a wrapped learner, e.g. a
TuningWrapper wrapped learner. It is also possible that below the
tuning wrapper, there are more CPOs. These can and will not be removed.

This function is complementary to \code{\link{getLearnerCPO}}.
}
\seealso{
Other CPOLearner related: 
\code{\link{CPOLearner}},
\code{\link{attachCPO}()},
\code{\link{getLearnerCPO}()}
}
\concept{CPOLearner related}
