% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublecaret.R
\name{internal\%>>\%}
\alias{internal\%>>\%}
\alias{internal\%<<\%}
\alias{internal\%<>>\%}
\alias{internal\%<<<\%}
\alias{internal\%>|\%}
\alias{internal\%|<\%}
\title{Internally Used \code{\%>>\%} Operators}
\usage{
"internal\%>>\%"(cpo1, cpo2)

"internal\%<<\%"(cpo2, cpo1)

"internal\%<>>\%"(cpo1, cpo2)

"internal\%<<<\%"(cpo2, cpo1)

"internal\%>|\%"(cpo1, cpo2)

"internal\%|<\%"(cpo2, cpo1)
}
\arguments{
\item{cpo1}{[\code{\link[base]{data.frame}} | \code{\link[mlr]{Task}} | \code{\link{CPO}} | \code{\link{CPOTrained}}]\cr
The source object.}

\item{cpo2}{[\code{\link{CPO}} | \code{\link{CPOTrained}} | \code{\link[mlr:makeLearner]{Learner}}]\cr
The sink object.}
}
\value{
[\code{\link[base]{data.frame}} | \code{\link[mlr]{Task}} | \code{\link{CPO}} | \code{\link{CPOTrained}}].
}
\description{
These functions are used internally as replacements of the
\code{\link{\%>>\%}} operators. This replacement is necessary
to enable right-associativity of some operators.
}
\keyword{internal}
