% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{is.inverter}
\alias{is.inverter}
\title{Check CPOInverter}
\usage{
is.inverter(x)
}
\arguments{
\item{x}{[any]\cr
The object to check.}
}
\value{
\code{TRUE} if \code{x} has class \code{\link{CPOInverter}}, \code{FALSE} otherwise.
}
\description{
Check whether the given object is a \code{\link{CPOInverter}} object.
}
\seealso{
Other inverter related: 
\code{\link{CPOTrained}},
\code{\link{NULLCPO}},
\code{\link{\%>>\%}()},
\code{\link{applyCPO}()},
\code{\link{as.list.CPO}},
\code{\link{clearRI}()},
\code{\link{getCPOClass}()},
\code{\link{getCPOName}()},
\code{\link{getCPOOperatingType}()},
\code{\link{getCPOPredictType}()},
\code{\link{getCPOProperties}()},
\code{\link{getCPOTrainedCPO}()},
\code{\link{getCPOTrainedCapability}()},
\code{\link{getCPOTrainedState}()},
\code{\link{makeCPOTrainedFromState}()},
\code{\link{pipeCPO}()},
\code{\link{print.CPOConstructor}()}
}
\concept{inverter related}
