% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NULLCPO.R
\name{is.nullcpo}
\alias{is.nullcpo}
\title{Check for NULLCPO}
\usage{
is.nullcpo(x)
}
\arguments{
\item{x}{[any]\cr
The object to check}
}
\value{
[\code{logical(1)}]. \code{TRUE} if \code{x} is a \code{NULLCPO}, \code{FALSE} otherwise.
}
\description{
Check whether the given object is a \code{\link{NULLCPO}}.
}
\seealso{
Other NULLCPO related: 
\code{\link{NULLCPO}},
\code{\link{nullToNullcpo}()},
\code{\link{nullcpoToNull}()}
}
\concept{NULLCPO related}
