% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{is.retrafo}
\alias{is.retrafo}
\title{Check CPORetrafo}
\usage{
is.retrafo(x)
}
\arguments{
\item{x}{[any]\cr
The object to check.}
}
\value{
\code{TRUE} if \code{x} has class \code{\link{CPORetrafo}}, \code{FALSE} otherwise.
}
\description{
Check whether the given object is a \code{\link{CPORetrafo}} object.
}
\seealso{
Other retrafo related: \code{\link{CPOTrained}},
  \code{\link{NULLCPO}}, \code{\link{\%>>\%}},
  \code{\link{applyCPO}}, \code{\link{as.list.CPO}},
  \code{\link{clearRI}}, \code{\link{getCPOClass}},
  \code{\link{getCPOName}},
  \code{\link{getCPOOperatingType}},
  \code{\link{getCPOPredictType}},
  \code{\link{getCPOProperties}},
  \code{\link{getCPOTrainedCPO}},
  \code{\link{getCPOTrainedCapability}},
  \code{\link{getCPOTrainedState}},
  \code{\link{makeCPOTrainedFromState}},
  \code{\link{pipeCPO}}, \code{\link{print.CPOConstructor}}
}
