% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCPO.R
\name{listCPO}
\alias{listCPO}
\title{List all Built-in CPOs}
\usage{
listCPO()
}
\description{
Return a \code{data.frame} with the columns \dQuote{name},
\dQuote{cponame}, \dQuote{category}, \dQuote{subcategory},
\dQuote{description}.

Categories and subcategories are:
\tabular{lll}{
  \bold{category} \tab \bold{subcategory}   \tab \bold{description}          \cr
  meta            \tab                      \tab CPO that acts on other CPOs \cr
  tools           \tab                      \tab                             \cr
  data            \tab general              \tab                             \cr
                  \tab general data preproc \tab                             \cr
                  \tab factor data preproc  \tab                             \cr
                  \tab numeric data preproc \tab                             \cr
                  \tab feature conversion   \tab                             \cr
                  \tab cleanup              \tab                             \cr
  featurefilter   \tab general              \tab fltr CPO with operation arg \cr
                  \tab specialised          \tab specific feat filter CPO    \cr
  imputation      \tab general              \tab imp CPO with operation arg  \cr
                  \tab specialised          \tab specific imputation CPO     \cr
  tools           \tab imputation           \tab                             \cr
}
}
