% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOTarget.R
\name{makeCPOTargetOp}
\alias{makeCPOTargetOp}
\title{Create a custom Target Operation CPO Constructor}
\usage{
makeCPOTargetOp(.cpo.name, ..., .par.set = NULL, .par.vals = list(),
  .dataformat = c("df.features", "split", "df.all", "task", "factor",
  "ordered", "numeric"), .dataformat.factor.with.ordered = TRUE,
  .data.dependent = TRUE, .trafo.type = c("trafo.returns.data",
  "trafo.returns.control", "stateless"), .export.params = TRUE,
  .properties.data = c("numerics", "factors", "ordered", "missings"),
  .properties.adding = character(0), .properties.needed = character(0),
  .properties.target = character(0), .type = c("cluster", "classif",
  "multilabel", "regr", "surv"), .type.out = .type,
  .predict.type = c(response = "response"), .packages = character(0),
  cpo.trafo, cpo.retrafo)
}
