% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NULLCPO.R
\name{nullcpoToNull}
\alias{nullcpoToNull}
\title{NULLCPO to NULL}
\usage{
nullcpoToNull(cpo)
}
\arguments{
\item{cpo}{[\code{\link{CPO}}]\cr
The cpo.}
}
\value{
[\code{\link{CPO}} | \code{NULL}]. \code{NULL} if \code{cpo} is \code{NULLCPO}, \code{cpo} otherwise.
}
\description{
Convert \code{\link{NULLCPO}} to \code{NULL}, leave other values intact.
}
\seealso{
Other NULLCPO related: 
\code{\link{NULLCPO}},
\code{\link{is.nullcpo}()},
\code{\link{nullToNullcpo}()}
}
\concept{NULLCPO related}
