% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NULLCPO.R
\name{nullToNullcpo}
\alias{nullToNullcpo}
\title{NULL to NULLCPO}
\usage{
nullToNullcpo(cpo)
}
\arguments{
\item{cpo}{[\code{\link{CPO}} | \code{NULL}]\cr
The CPO.}
}
\value{
[\code{\link{CPO}}]. \code{\link{NULLCPO}} if \code{cpo} is \code{NULL}, \code{cpo} otherwise.
}
\description{
Convert \code{NULL} to \code{\link{NULLCPO}}, leave other values intact.
}
\seealso{
Other NULLCPO related: \code{\link{NULLCPO}},
  \code{\link{is.nullcpo}}, \code{\link{nullcpoToNull}}
}
\concept{NULLCPO related}
