% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{setCPOId}
\alias{setCPOId}
\title{Set the ID of a CPO object.}
\usage{
setCPOId(cpo, id)
}
\arguments{
\item{cpo}{[\code{CPO}]\cr
The \code{CPO} object to modify.}

\item{id}{[\code{character(1)} | \code{NULL}]\cr
The ID. If this is \code{NULL}, the ID is set to the
default for the CPO at hand.}
}
\value{
[\code{CPO}] the CPO with modified id.
}
\description{
Setting the ID of a CPO to a value will prefix all its
parameter names with this ID. This makes it possible to
compose CPOs that have clashing parameter names.
}
\seealso{
Other CPO: \code{\link{CPOImputer}}, \code{\link{\%>>\%}},
  \code{\link{applyCPO}}, \code{\link{attachCPO}},
  \code{\link{cpoApplyFun}}, \code{\link{cpoAsNumeric}},
  \code{\link{cpoCase}}, \code{\link{cpoCollapseFact}},
  \code{\link{cpoDropConstants}},
  \code{\link{cpoDummyEncode}},
  \code{\link{cpoFilterFeatures}},
  \code{\link{cpoFixFactors}},
  \code{\link{cpoImpactEncodeClassif}},
  \code{\link{cpoImpactEncodeRegr}},
  \code{\link{cpoImpute}},
  \code{\link{cpoMissingIndicators}},
  \code{\link{cpoModelMatrix}}, \code{\link{cpoMultiplex}},
  \code{\link{cpoPca}}, \code{\link{cpoProbEncode}},
  \code{\link{cpoQuantileBinNumerics}},
  \code{\link{cpoScaleMaxAbs}},
  \code{\link{cpoScaleRange}}, \code{\link{cpoScale}},
  \code{\link{cpoSelect}}, \code{\link{cpoSpatialSign}},
  \code{\link{cpoWrap}}, \code{\link{getCPOAffect}},
  \code{\link{getCPOId}}, \code{\link{getCPOName}},
  \code{\link{getRetrafoState}},
  \code{\link{makeCPOExtended}}, \code{\link{makeCPO}},
  \code{\link{makeRetrafoFromState}}, \code{\link{pipeCPO}}
}
