% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getCPOName}
\alias{getCPOName}
\title{Get the CPO object's Name}
\usage{
getCPOName(cpo)
}
\arguments{
\item{cpo}{[\code{\link{CPO}}]\cr
The CPO object.}
}
\description{
Return the name given at creation as \dQuote{.cpo.name} to the
CPO creator. If the CPO object has an ID, it will be appended.
}
\seealso{
Other CPO: \code{\link{CPOImputer}}, \code{\link{CPO}},
  \code{\link{\%>>\%}}, \code{\link{applyCPO}},
  \code{\link{attachCPO}}, \code{\link{cpoApplyFun}},
  \code{\link{cpoAsNumeric}}, \code{\link{cpoCase}},
  \code{\link{cpoCollapseFact}},
  \code{\link{cpoDropConstants}},
  \code{\link{cpoDummyEncode}},
  \code{\link{cpoFilterFeatures}},
  \code{\link{cpoFixFactors}},
  \code{\link{cpoImpactEncodeClassif}},
  \code{\link{cpoImpactEncodeRegr}},
  \code{\link{cpoImpute}},
  \code{\link{cpoMissingIndicators}},
  \code{\link{cpoModelMatrix}}, \code{\link{cpoMultiplex}},
  \code{\link{cpoPca}}, \code{\link{cpoProbEncode}},
  \code{\link{cpoQuantileBinNumerics}},
  \code{\link{cpoScaleMaxAbs}},
  \code{\link{cpoScaleRange}}, \code{\link{cpoScale}},
  \code{\link{cpoSelect}}, \code{\link{cpoSpatialSign}},
  \code{\link{cpoWrap}}, \code{\link{getCPOAffect}},
  \code{\link{getCPOId}}, \code{\link{getRetrafoState}},
  \code{\link{inverter<-}}, \code{\link{inverter}},
  \code{\link{makeCPOExtended}}, \code{\link{makeCPO}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{pipeCPO}}, \code{\link{retrafo<-}},
  \code{\link{retrafo}}, \code{\link{setCPOId}}
}
