% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NULLCPO.R
\docType{data}
\name{NULLCPO}
\alias{NULLCPO}
\title{CPO composition neutral element}
\format{An object of class \code{NULLCPO} (inherits from \code{CPOPrimitive}, \code{CPORetrafo}, \code{CPOInverter}, \code{CPOTrained}, \code{CPO}) of length 0.}
\usage{
NULLCPO
}
\description{
\code{NULLCPO} is the neutral element of \code{CPO} and \code{CPOTrained} composition. It is returned
when \code{\link{trafo}} and \code{\link{retrafo}} are applied to an object that has no retrafo or inverter
associated with it.
}
\seealso{
Other NULLCPO: \code{\link{is.nullcpo}},
  \code{\link{nullToNullcpo}}, \code{\link{nullcpoToNull}}
}
\keyword{datasets}
